Structure tests (#222)
* added tests module

* put unittests for TensorNetwork into separate file

* put serialization tests in separate file

* put factorization method tests in separate file

* changes to tensornetwork tests

* set jaxlib version

* no longer support python3.5

* added jax version

* removed todos added tests for public methods

* unused variable
