+++ .travis.yml
@@ -1,7 +1,6 @@
 dist: xenial 
 language: python
 python:
-  - "3.5"
   - "3.6"
   - "3.7"
 # command to install dependencies
