+++ requirements_travis.txt
@@ -1,5 +1,5 @@
-jax
-jaxlib
+jax>=0.1.0
+jaxlib>=0.1.27
 pytype>=2019.06.21
 pytest
 torch>=1.1
